soc: hobot: axi_monitor: feat [x5-65] Add automatic power on/off and clk on/off for qos read/write
Solve the problem that the priority register is restored to the default
value when each power domain powered off and then on.And the qos priority of each submodule can be set even when the power domain is powered off.Save read/write priority value and mode value when qos is initialized.
Turned on and off according to the clk required by qos, when the
priority of qos is operated or read.

Signed-off-by: Hualun Dong <hualun.dong@horizon.cc>
Change-Id: I4a237f5c2206fb85a40f1bc861a9f96baabfa4c6