net: stmmac: fix [X5-835] disable Split Header (SPH) for Sunrise5 platforms.
Based on DesignWare Ethernet QoS datasheet, we are seeing the limitation
of Split Header (SPH) feature is not supported for Ipv4 fragmented packet.
This SPH limitation will cause ping failure when the packets size exceed
the MTU size. For example, the issue happens once the basic ping packet
size is larger than the configured MTU size and the data is lost inside
the fragmented packet, replaced by zeros/corrupted values, and leads to
ping fail.

So, disable the Split Header for Sunrise5 platforms.

Change-Id: I3acfc67c0f9f06db0f7ceac4262c555b3e4af347
Signed-off-by: jianghe.xu <jianghe.xu@horizon.cc>