staging: ion: refactor [X5-2093] Change dts compatible matching HEAP_TYPE
Summary:
Before this patch, "ion-pool" is configured to ION_HEAP_TYPE_CARVEOUT and
"ion-carveout" is configured to ION_HEAP_TYPE_CMA_RESERVED which makes
the information regarding "ion-pool" is shown in debugfs under the name of
"carveout" and information regarding "ion-carveout" is shown in debugfs under
the name of "cma_reserved". Switch the two to make debugfs information more
intuitive.

The result of this is the actual physical address of the two heaps
are switched as well.

Signed-off-by: dinggao.pan <dinggao.pan@d-robotics.cc>
Change-Id: I835912155fda4e20604f85f83cf87a20ae8300aa